Bifurcated the Actors into The People vs. The State.
