+++ src/People/SuspectGenome.php
@@ -10,13 +10,16 @@
  * This file is licensed under the MIT License.
  */
 
-namespace HopeSeekr\PrisonersDilemma;
+namespace HopeSeekr\PrisonersDilemma\People;
 
-interface CriminalCode
+class SuspectGenome
 {
-    public function isValidSentence(int $sentenceId): bool;
+    public array $actions = [];
+    public int $actionWeight = 0;
 
-    public function getSentence($sentenceId): float;
-
-    public function getMaxSentence(): float;
+    public function rememberAction(int $actionId, string $actionDescription)
+    {
+        // @todo: Should probably add error handling if the action doesn't actually exist.
+        $this->actions[$actionId] = $actionDescription;
+    }
 }
